% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{arabidopsis}
\alias{arabidopsis}
\title{The Arabidopsis RNA-Seq Dataset}
\format{A 21567 by 12 matrix of RNA-Seq read frequencies.}
\usage{
data(arabidopsis)
}
\description{
The goal of this study is to identify the targets of RVE8, a MYB-like transcription factor involved in the circadian clock in Arabidopsis. Analysis of 7 days old rve8-1 RVE8::RVE8:GR and rve8-1 seedlings treated with dexamethasone or mock identified genes responsive to RVE8 induction.The RVE8 up-regulated genes are enriched for evening-phased genes while the down-regulated genes are enriched for a morning phase. This study reveals that RVE8 is a master regulator of circadian gene expression in Arabidopsis. See References below.
}
\details{
Overview Design: Transgenic line rve8-1 RVE8::RVE8:GR and rve8-1 treated with DEX or mock with three biological replicates each, 12 samples in total. See References below.
}
\references{
See \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE38879} for more details.
}
\keyword{#}
\keyword{,10:12]}
\keyword{,1:3]}
\keyword{,4:6]}
\keyword{,7:9]}
\keyword{=}
\keyword{RVE8:GR_DEX}
\keyword{RVE8:GR_mock}
\keyword{RVE8_GR_DEX}
\keyword{RVE8_GR_mock}
\keyword{arabidopsis[}
\keyword{data(arabidopsis)}
\keyword{datasets}
\keyword{dim(arabidopsis)}
\keyword{group}
\keyword{head(arabidopsis)}
\keyword{library(SeqDisp)}
\keyword{only:}
\keyword{rve8_DEX}
\keyword{rve8_mock}

