% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cali.sigma.plot}
\alias{cali.sigma.plot}
\title{Plotting function for quantifying noise with calibration interval in ggplot2}
\usage{
cali.sigma.plot(real.data.sigma.obj, sim.sigma.obj, sigma.v = c(0.5, 0.7, 0.8,
  0.9, 1, 1.1, 1.2, 1.5), cali.sigma = 1, left.ci = 0.8, right.ci = 1.2,
  title = NULL)
}
\arguments{
\item{real.data.sigma.obj}{an object from \code{quantify.noise} where the second evaluate is TRUE}

\item{sim.sigma.obj}{an object from \code{quantify.noise} where the third evaluate is TRUE}

\item{sigma.v}{a vector of true sigma (to be plotted on the x-axis)}

\item{cali.sigma}{the calibrated sigma estimate (need to see the plot and then determine this value)}

\item{left.ci}{the left bound for the calibration interval (need to see the plot and then determine this value)}

\item{right.ci}{the right bound for the calibration interval (need to see the plot and then determine this value)}

\item{title}{main title of the plot}
}
\value{
a ggplot2 object
}
\description{
Plotting function for quantifying noise with calibration interval in ggplot2
}
\author{
Gu Mi
}

