% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimate.dispersion.var.edgeR}
\alias{estimate.dispersion.var.edgeR}
\title{Estimate extra variation in dispersion after a dispersion model has been fit (for edgeR)}
\usage{
estimate.dispersion.var.edgeR(nb.data, dispersion, x, subset.rows = 1:m,
  lower = 0, upper = 200, hessian = FALSE)
}
\arguments{
\item{nb.data}{output from \code{link{prepare.nb.data}}}

\item{dispersion}{dispersion matrix of dimension m-by-n (edgeR)}

\item{x}{model matrix, should be the same one used to estimate \code{dipserison}}

\item{subset.rows}{the row indices for subsetting the read count matrix}

\item{lower}{lower bound in \code{optimize}}

\item{upper}{upper bound in \code{optimize}}

\item{hessian}{(logical) whether to calculate the Hessian matrix}
}
\value{
an object
}
\description{
Estimate extra variation in dispersion after a dispersion model has been fit (for edgeR)
}
\details{
The fitted dispersion models may come from the edgeR packages. This function are used for both simulation
(see sim.dr) and real RNA-Seq data studies.
}
\author{
Yanming Di, Gu Mi
}

