% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit.nb.regression.1}
\alias{fit.nb.regression.1}
\title{Find mle of the dispersion and the regression coefficients in a NB regression model}
\usage{
fit.nb.regression.1(y, s, x, beta0 = rep(NA, dim(x)[2]), kappa = NA)
}
\arguments{
\item{y}{}

\item{s}{}

\item{x}{}

\item{beta0}{}

\item{kappa}{}
}
\value{
a list
}
\description{
Find mle of the dispersion and the regression coefficients in a NB regression model
}
\author{
Yanming Di
}

