% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{fruit.fly}
\alias{fruit.fly}
\title{The Drosophila melanogaster (common fruit fly) RNA-Seq Dataset}
\format{A 14470 by 7 matrix of RNA-Seq read frequencies.}
\usage{
data(fruit.fly)
}
\description{
See References below.
}
\details{
This dataset is directly obtained from the \pkg{pasilla} Bioconductor package, which provides per-exon and per-gene read counts computed for selected genes in Brooks, Yang, Duff, Hansen, Park, et al. It can also be accessed from \code{data(pasillaGenes)} once \pkg{pasilla} is loaded into session.
}
\references{
Brooks et al. Conservation of an RNA regulatory map between Drosophila and mammals. Genome Research, 2010.
See the vignette of the \pkg{pasilla} package for more details.
}
\keyword{datasets}

