% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genewise.hoa}
\alias{genewise.hoa}
\title{Implement genewise dispersion approach with LRT-HOA test}
\usage{
genewise.hoa(nb.data, grp.ids, grp1, grp2, R = 100, print.level = 1)
}
\arguments{
\item{nb.data}{}

\item{grp.ids}{}

\item{grp1}{}

\item{grp2}{}

\item{R}{}

\item{print.level}{}
}
\value{
a data frame where each row is from hoa.1u() function output
}
\description{
Implement genewise dispersion approach with LRT-HOA test
}
\author{
Yanming Di
}

