% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hoa.1u}
\alias{hoa.1u}
\title{HOA test for regresssion coefficeint in a NB regression model}
\usage{
hoa.1u(y, s, x, beta0, alternative = "two.sided", R = 100,
  print.level = 0)
}
\arguments{
\item{y}{}

\item{s}{}

\item{x}{}

\item{beta0}{}

\item{alternative}{}

\item{R}{}

\item{print.level}{}
}
\value{
a data frame
}
\description{
HOA test for regresssion coefficeint in a single NB regression model
}
\author{
Yanming Di
}
\keyword{internal}

