% Generated by roxygen2 (4.0.1): do not edit by hand
\name{il.prior.var.1}
\alias{il.prior.var.1}
\title{Compute integrated likelihood of the variance of the
prior distribution of log dispersion}
\usage{
il.prior.var.1(prior.var, prior.log.phi, y, s, x, beta0 = rep(NA, dim(x)[2]))
}
\arguments{
\item{prior.var}{a number, the variance of the prior distribution of log(phi) --> sigma^2}

\item{prior.log.phi}{a number or an n-vector, the mean(s) of the prior distribution of log(phi)  --> theta0}

\item{y}{a n-vector of NB counts.}

\item{s}{a n-vector of library sizes.}

\item{x}{a n by p design matrix.}

\item{beta0}{a p-vector specifying the known and unknown
components of beta, the regression coefficients. NA values
indicate unknown components and non-NA values specify the values
of the known components. The default is that all components of
beta are unknown.}
}
\value{
a list containing the following components:
  \item{l}{the integrated likelihood of \code{prior.var}}
  \item{epsilon.map}{a scalar, MAP estimate of epsilon = log.phi - prior.log.phi}
  \item{log.phi.map}{a scalar or a vector, MAP estimate of log.phi}
}
\description{
This function finds the MAP estimator of log(phi) and computes
the integrated likelihood (IL) of the variance of the prior
distribution on log(phi) by Laplace approximation.
}
\details{
Note that the components of log(phi.prior) can be different, but
it is assumed that

epsilon = log(phi) - prior.log.phi

is the same for all components and epsilon has a (prior) distribution

N(0, prior.var).

Given phi, the components of y follow a NB distribution with
means mu = s exp(x' beta) and dispersion phi.

The integrated likelihood of \code{prior.var} is defined as the integration of

Lp(epsilon; y) Prior(epsilon)

over epsilon, where Lp(epsilon; y) is the profile likelihood of epsilon.

For given epsilon, we can determine the value of phi=phi(epsilon),
and find the MLE of beta. The profile likelihood of epsilon is the
likelihood of (epsilon, beta) computed at epsilon and the MLE of
beta given epsilon: l(phi(epsilon), beta_MLE(phi)).
}
\author{
Yanming Di
}

