% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mddata}
\alias{mddata}
\title{Preparing data frames for producing mean-dispersion plot using ggplot2 for several NB dispersion models}
\usage{
mddata(counts, x, model = NULL, sc = 1)
}
\arguments{
\item{counts}{RNA-Seq raw read count matrix}

\item{x}{design matrix}

\item{model}{name of the NB dispersion model, including "NBP", "NBQ", "NBS", "STEP", "Common",
"Tagwise-Common", "Tagwise-Trend" and "Trended"}

\item{sc}{a scaling factor to MOM estimator phi, for an overall up-/down-shift of the dispersions (default = 1)}
}
\value{
depending on the model, this function returns coordinates of the scatter plot and names of the dispersion models
}
\description{
Preparing data frames for producing mean-dispersion plot using ggplot2 for several NB dispersion models
}
\details{
We temporarily use method-of-moments (MOM) estimators for each gene in the scatter plot.
}
\author{
Gu Mi
}
\keyword{internal}

