% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{mouse}
\alias{mouse}
\title{The Mus Musculus (Mouse) RNA-Seq Dataset}
\format{A 39178 by 9 matrix of RNA-Seq read frequencies.}
\usage{
data(mouse)
}
\description{
The goals of this study were to compare signals from competent and abnormal human embryos impacted differently on the expression of endometrial receptivity genes in mouse uteri using transcriptome profiling (RNA-seq). See References below.
}
\details{
Overview Design: Supernatant from competent and abnormal human embryos were flushed into the uterine horns of 25d old C57BL/6 female mice. Uterine mRNA profiles of 21-day old mice were generated by deep sequencing, in triplicate, using Illumina HiSeq 2000 platform (HiSeq). See References below.
}
\examples{
library(SeqDisp)
data(mouse)
head(mouse)
dim(mouse)

# Control group only:
control = mouse[ ,seq(1,3)]

# DCECM group only:
DCECM = mouse[ ,seq(4,6)]

# AECM group only:
AECM = mouse[ ,seq(7,9)]
}
\references{
See \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE47019} for more details.
Web link: \url{http://www.nature.com/srep/2014/140206/srep03894/full/srep03894.html}.
}
\keyword{datasets}

