% Generated by roxygen2 (4.0.2): do not edit by hand
\name{n.tp.plot}
\alias{n.tp.plot}
\title{Plot the top N genes vs. True Positve Rate (TPR) plot using ggplot2 (with other information reported)}
\usage{
n.tp.plot(test.data.list, groupName = "grp", predName = "pvals",
  legend = TRUE, legend.just = c(0, 1), N = 500, panel.label = "A",
  title = "Plot Title")
}
\arguments{
\item{test.data.list}{meta-data of the quantities for each dispersion model, as prepared in sim.pwr.plot()}

\item{groupName}{leave as default "grp"; c.f. sim.pwr.plot()}

\item{predName}{leave as default "pvals"; c.f. sim.pwr.plot()}

\item{legend}{whether to put a legend on a single plot}

\item{legend.just}{the legend.justification argument in ggplot2 controling the position of the legend}

\item{N}{the number of top genes selected}

\item{panel.label}{the label on the lower-right corner of each plot (for journal figure legends)}

\item{title}{the plot title for a single panel}
}
\value{
a list of a ggplot2 object and relevant information.
}
\description{
Plot the top N genes vs. True Positve Rate (TPR) plot using ggplot2 (with other information reported)
}
\details{
This function produces a single ggplot2 object, and is made invisible to end-users. For multiple plots, see the
\code{pl.n.tp} function.
}
\author{
Gu Mi
}
\keyword{internal}

