% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optimize.il.prior.var}
\alias{optimize.il.prior.var}
\title{Estimate extra variation in dispersion after a dispersion model has been fit}
\usage{
optimize.il.prior.var(y, s, x, log.phi, lower = 0, upper = 200,
  hessian = FALSE)
}
\arguments{
\item{y}{read count matrix}

\item{s}{library size}

\item{x}{design matrix}

\item{log.phi}{pre-estimate dispersion matrix}

\item{lower}{lower bound in \code{optimize}}

\item{upper}{upper bound in \code{optimize}}

\item{hessian}{(logical) whether to calculate the Hessian matrix}
}
\value{
an object
}
\description{
This function is similar to estimate.dispersion.var, except that it accepts log.phi directly
}
\author{
Yanming Di
}

