% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pl.ma}
\alias{pl.ma}
\title{Modified MA Plot for RNA-Seq Data Analysis}
\usage{
pl.ma(counts, x, model = NULL, FDR = NULL, top = NULL, gradient = FALSE,
  title = paste0("MA Plot: ", model, " Model"))
}
\arguments{
\item{counts}{raw RNA-Seq read counts}

\item{x}{design matrix}

\item{model}{NB dispersion model name (used for testing DE). Specify "NBP", "NBQ", "NB2", "Genewise-HOA" (in the NBPSeq package),
"Common", "Genewise", "Trended", "Tagwise-Common" or "Tagwise-Trend" (in the edgeR package)}

\item{FDR}{threshold of false discovery rate (default = 0.1) to call a gene significant}

\item{top}{a scalar showing the number of most significant DE genes}

\item{gradient}{(logical) whether to use color gradient to show FDR for DE genes (if \code{FDR=TRUE}), or show estimated
dispersions for the genewise model (if \code{top=TRUE}). If \code{FALSE}, DE genes will be shown in solid red.}

\item{title}{title of the plot}
}
\value{
a ggplot2 object
}
\description{
Typical MA plot would not have the significance. This modified function introduces the significance
used in Volcano plot into MA plot for the selected DE genes (by FDR), to better reveal the relationship between test significance
and gene expression levels (fold change). We use color gradient to show the magnitude of FDR (red: most significant;
blue: significant but close to the FDR threshold). In addition, for the genewise approach, we use color gradient to show the dispersion estimates for the top selected DE genes. For other approaches, we plot in red the top-selected genes (by the \code{top} argument).
}
\examples{
library(SeqDisp)

data(mouse)
counts = mouse[1:5000,seq(1,6)]  # subset to 5000 genes and two groups
grp.ids = as.factor(c(1,1,1,2,2,2))
x = model.matrix(~grp.ids)

# select DE genes by FDR threshold
map.nbp.fdr = pl.ma(counts, x, model="NBP", FDR=0.1, gradient=TRUE)
map.gen.fdr = pl.ma(counts, x, model="Genewise", FDR=0.1, gradient=TRUE)
multiplot(map.nbp.fdr, map.gen.fdr, cols=2, layout=matrix(seq(1,2), nr=1, byrow=TRUE))

# select top DE genes
map.nbp.top = pl.ma(counts, x, model="NBP", top=100)
map.gen.top = pl.ma(counts, x, model="Genewise", top=100)
multiplot(map.nbp.top, map.gen.top, cols=2, layout=matrix(seq(1,2), nr=1, byrow=TRUE))
}
\author{
Gu Mi
}

