% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl.md}
\alias{pl.md}
\title{Plot mean-dispersion scatter plot (log-log scale) for several NB dispersion models using ggplot2}
\usage{
pl.md(model.vec, counts, x, sc = 1, title = NULL, data.note = NULL)
}
\arguments{
\item{model.vec}{a vector of NB dispersion models where the fitted curves will be superimposed on the scatter plot.
Specify: "NBP", "NBQ", "NBS", "STEP", "Common", "Tagwise-Common", "Tagwise-Trend" or "Trended"}

\item{counts}{RNA-Seq raw read count matrix}

\item{x}{design matrix}

\item{sc}{a scaling factor to MOM estimator phi, for an overall up-/down-shift of the dispersions (default = 1)}

\item{title}{main title of the plot}

\item{data.note}{note of the dataset (put in the lower-left corner of the plot)}
}
\value{
depending on the model, this function returns coordinates of the scatter plot and names of the dispersion models
}
\description{
Plot mean-dispersion scatter plot (log-log scale) for several NB dispersion models using ggplot2
}
\details{
We temporarily use method-of-moments (MOM) estimators for each gene in the scatter plot, and superimpose
fitted curves for the dispersion models specified in \code{model.vec}.
}
\author{
Gu Mi
}

