% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pl.n.tp}
\alias{pl.n.tp}
\title{Plot the top N genes vs. True Positve Rate (TPR) plot using ggplot2}
\usage{
pl.n.tp(dt = NULL, sub.col = 1:6, grp.ids = c(rep(1, length(sub.col)/2),
  rep(2, length(sub.col)/2)), disp.vec = NULL, init.fit = TRUE, m = 5000,
  n = 6, s = 1e+07, de.vec = 0.2, fc.vec = log(3), noise.vec = 0,
  legend.just = c(0, 1), N = 500, seed = 539, out.path = getwd(),
  model.vec = NULL)
}
\arguments{
\item{dt}{the full dataset to be loaded and used as a template for the simulation}

\item{sub.col}{column indices used to subset the full dataset}

\item{grp.ids}{group ids to distinguish between the two groups}

\item{disp.vec}{the dispersion types to be simulated. Specify "Linear", "Quadratic" or "Non-parametric"}

\item{init.fit}{(logical) whether to estimate the dispersion model and save as a template (for later use)}

\item{m}{number of genes (default = 5000 genes) subsetted for actual analysis}

\item{n}{number of samples (default = 6 samples)}

\item{s}{library size (default = 1e7, 10 million reads per library)}

\item{de.vec}{a vector of percentages of DE genes (default = 0.2)}

\item{fc.vec}{a vector of log fold-changes, i.e. the value of beta2 (default = log(3.0), for FC = 3.0).
If a vector of fold changes, will call \code{sim.data.prepare};
If NULL, will call \code{sim.data.prepare2}.}

\item{noise.vec}{a vector of random normal(0, sigma) noise to phi: phi * exp(rnorm(m, 0, sigma)).
If any of the noise.vec elements equals 0 (default), data without noise are simulated}

\item{legend.just}{the legend.justification argument in ggplot2 controling the position of the legend.}

\item{N}{the number of top genes selected}

\item{seed}{random generator seed for reproducible research}

\item{out.path}{directory path where the intermediate outputs are stored (and then fetched)}

\item{model.vec}{a vector of NB dispersion models to be compared for power-robustness, including:
"Common", "NBP", "NBQ", "NBS", "Trended", "Genewise", "Tagwise-Common", "Tagwise-Trend", "QL", "QLShrink" and "QLSpline"}
}
\value{
a list of two component: (1) information for ggplot2 plotting; (2) useful quantities in calculating TPR and FDR.
Also, this function output .txt files for future use.
}
\description{
Plot the top N genes vs. True Positve Rate (TPR) plot using ggplot2
}
\details{
This function calls \code{sim.data.prepare} to generate simulated read counts with different dispersion models,
varying percentages of DE genes, log fold changes and random normal noise levels. This function calls \code{sim.pwr.plot}
to generate a ggplot2 object for a single scenario (DE percentage, log FC and noise level combination) and then loop over
all considered scearios (panels of plots). This function also output .txt files (for each panel) the N, TP, p-value and
q-values for each of the models considered (as specified in the \code{model.vec} argument).
}
\author{
Gu Mi
}

