% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pl.voc}
\alias{pl.voc}
\title{Modified Volcano Plot for RNA-Seq Data Analysis}
\usage{
pl.voc(counts, x, model = NULL, title = paste0("Volcano Plot: ", model,
  " Model"))
}
\arguments{
\item{counts}{raw RNA-Seq read counts}

\item{x}{design matrix}

\item{model}{NB dispersion model name (used for testing DE). Specify "NBP" (in the NBPSeq package), "Common", "Genewise", "Trended",
"Tagwise-Common" or "Tagwise-Trend" (in the edgeR package)}

\item{title}{title of the plot}
}
\value{
a ggplot2 object
}
\description{
We use color gradient to show the magnitude of significance.
}
\examples{
library(SeqDisp)

data(mouse)

counts = mouse[1:5000,seq(1,6)]  # subset to 5000 genes and two groups
grp.ids = as.factor(c(1,1,1,2,2,2))
x = model.matrix(~grp.ids)

vocplot.nbp = pl.ma(counts, x, model="NBP")
vocplot.gen = pl.ma(counts, x, model="Genewise")

multiplot(vocplot.nbp, vocplot.gen, cols=2, layout=matrix(seq(1,2), nr=1, byrow=TRUE))
}
\author{
Gu Mi
}

