% Generated by roxygen2 (4.0.2): do not edit by hand
\name{poly.gamma.reg}
\alias{poly.gamma.reg}
\title{Fit polynomial gamma log-linear regression with linear, quadratic or cubic terms}
\usage{
poly.gamma.reg(counts, x, model = "linear")
}
\arguments{
\item{counts}{read counts matrix}

\item{x}{design matrix}

\item{model}{choose from "linear", "quadratic" or "cubic"}
}
\value{
a list showing the significance, R^2 of each of the fitted polynomial regression
}
\description{
Fit polynomial gamma log-linear regression with linear, quadratic or cubic terms
}
\author{
Gu Mi
}

