% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prec.plot}
\alias{prec.plot}
\title{Plot the Precision-Recall plot using ggplot2 (with other information and F-measures reported)}
\usage{
prec.plot(test.data.list, beta = beta, groupName = "grp",
  predName = "pvals", legend = TRUE, legend.just = c(0, 1),
  x.limit = c(0, 1), panel.label = "A", title = "Plot Title")
}
\arguments{
\item{test.data.list}{meta-data of the quantities for each dispersion model, as prepared in sim.pwr.plot()}

\item{beta}{a non-negative real value for calculating the general F_beta measure (default = 1: recall and precision are evenly weighted)}

\item{groupName}{leave as default "grp"; c.f. sim.pwr.plot()}

\item{predName}{leave as default "pvals"; c.f. sim.pwr.plot()}

\item{legend}{whether to put a legend on a single plot}

\item{legend.just}{the legend.justification argument in ggplot2 controling the position of the legend}

\item{x.limit}{specify the plotting range for the x-axis (TPR), for better visualization}

\item{panel.label}{the label on the lower-right corner of each plot (for journal figure legends)}

\item{title}{the plot title for a single panel}
}
\value{
a list of a ggplot2 object, relevant information and F-measures.
}
\description{
Plot the Precision-Recall plot using ggplot2 (with other information and F-measures reported)
}
\details{
This function produces a single ggplot2 object, and is made invisible to end-users. For multiple plots, see the
\code{pl.prec} function.
}
\author{
Gu Mi
}
\keyword{internal}

