% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pwrdata}
\alias{pwrdata}
\title{Preparing quantities for TPR vs. FDR plot, or Precision-Recall plot}
\usage{
pwrdata(grp, pred, type = "TPR-FDR", N = 500, beta = beta)
}
\arguments{
\item{grp}{labels classifying subject status (true DE labels)}

\item{pred}{values of each observation (DE test p-values)}

\item{type}{which type of plot to draw? Either "TPR-FDR", "N-TP" or "PREC"}

\item{N}{the number of top genes selected}

\item{beta}{a non-negative real value for calculating the general F_beta measure (default = 1: recall and precision are evenly weighted)}
}
\value{
List with the common component:

pwr: data.frame with x and y co-ordinates of plot (TPR and FDR, or recall and precision),
     and sorted p-values, q-values by B-H adjustment and F-measures (if PR-curve is desired).
}
\description{
Preparing quantities for TPR vs. FDR plot, or Precision-Recall plot
}
\details{
This function produces x and y co-ordinates for either type of plots.
}
\author{
Gu Mi
}
\keyword{internal}

