% Generated by roxygen2 (4.0.2): do not edit by hand
\name{quantify.noise}
\alias{quantify.noise}
\title{Prepare NB dispersions for NBQ dispersion models, and calculate "sigma" in real and simulated datasets}
\usage{
quantify.noise(dt = NULL, sigma.v = c(0.5, 0.7, 0.8, 0.9, 1, 1.1, 1.2, 1.5),
  m = 5000, sub.col = NULL, grp.ids = c(rep(1, length(sub.col)/2), rep(2,
  length(sub.col)/2)), evaluate = c(TRUE, FALSE, FALSE), method = "ML",
  hessian = FALSE, seed = 539, path.o = getwd())
}
\arguments{
\item{dt}{name of the dataset}

\item{sigma.v}{a vector of true sigma}

\item{m}{number of genes to subset}

\item{sub.col}{indices for subsetting the columns of dt}

\item{grp.ids}{group ids to distinguish between the two groups}

\item{evaluate}{a vector of logicals for evaluating underlying dispersion, real data sigma, and simulated data sigma's}

\item{method}{estimation method for dispersion to be used for quantify residual dispersion variation, either "ML" or "MAPL"
(default is "ML")}

\item{hessian}{whether to calculate the hessian matrix for getting standard error of the estimated sigma (default is FALSE)}

\item{seed}{random generator seed for sampling a subset of genes}

\item{path.o}{output directory}
}
\value{
many saved intermediate files used for quantifying noise
}
\description{
Prepare NB dispersions for NBQ dispersion models, and calculate "sigma" in real and simulated datasets
}
\details{
The dispersion results are saved as RData files, which will be used in subsequent steps to calculate

  log.phi.mean = cbind(1, z, z^2) %*% dispersion$model$par

e.g. the "dispersion" object
}
\author{
Gu Mi
}

