% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{SeqDisp-package}
\alias{SeqDisp}
\alias{SeqDisp-package}
\title{Evaluations of Negative Binomial Dispersion Methods for RNA Sequencing Data}
\description{
An R package for evaluating the negative binomial dispersions methods commonly used in RNA-Seq data analysis
}
\details{
Details of the package
}
\author{
Yanming Di <diy@stat.oregonstate.edu>, Gu Mi <neo.migu@gmail.com>, Daniel W. Schafer

Maintainer: Gu Mi <https://github.com/gu-mi>
}
\references{
See \url{https://github.com/gu-mi/SeqDisp/wiki/} for more details.
}
\keyword{package}

