% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.data.prepare2}
\alias{sim.data.prepare2}
\title{Preparing data for the power-robustness simulations}
\usage{
sim.data.prepare2(dt = NULL, sub.col = 1:6, grp.ids = c(rep(1,
  length(sub.col)/2), rep(2, length(sub.col)/2)), disp = "Quadratic",
  method = "MAPL", init.fit = TRUE, m = 5000, n = 6, s = 1e+07,
  perc.de = 0.2, sigma = 0, seed = 539, out.path = getwd())
}
\arguments{
\item{dt}{the full dataset to be loaded and used as a template for the simulation}

\item{sub.col}{column indices used to subset the full dataset}

\item{grp.ids}{group ids to distinguish between the two groups}

\item{disp}{the dispersion type to be simulated. Specify "Linear", "Quadratic" or "Non-parametric"}

\item{method}{estimation method for dispersion, either "ML" or "MAPL" (default is "MAPL")}

\item{init.fit}{(logical) whether to estimate the dispersion model and save as a template (for later use)}

\item{m}{number of genes (default = 5000 genes) subsetted for actual analysis}

\item{n}{number of samples (default = 6 samples)}

\item{s}{library size (default = 1e7, 10 million reads per library)}

\item{perc.de}{percentage of DE genes (default = 0.2)}

\item{sigma}{random normal(0, sigma) noise to phi: phi * exp(rnorm(m, 0, sigma)).
If sigma = 0 (default), data without noise are simulated}

\item{seed}{random generator seed for reproducible research}

\item{out.path}{directory path where the intermediate outputs are stored (and then fetched)}
}
\value{
a list of read counts ("y"), design matrix ("x"), library size ("s"), indices of DE genes ("idx.DE"),
percentage of DE genes ("perc.de"), number of genes ("m"), number of samples ("n"), dispersion ("phi"), and estimated relative
mean frequency ("pi")
}
\description{
Preparing data for the power-robustness simulations
}
\details{
Currently we can simulate three types of dispersion models: "Linear" (NBP), "Quadratic" (NBQ) or "Non-parametric" (Trended).
This function differs from \code{sim.data.prepare} by using fold changes from real data estimates instead of specifying a scalar.
}
\author{
Gu Mi
}

