% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.dr}
\alias{sim.dr}
\title{Estimate sigma on simulated datasets}
\usage{
sim.dr(m, n, x, s, mu, sigma, seed, log.phi.mean = NULL, model = NULL,
  method = "ML", hessian = FALSE, refit = TRUE)
}
\arguments{
\item{m}{number of genes}

\item{n}{number of samples}

\item{x}{design matrix}

\item{s}{library size}

\item{mu}{mean level}

\item{sigma}{a vector of true sigma values}

\item{seed}{random number generator seed (for N(0,sigma^2) noise)}

\item{log.phi.mean}{a m-by-n matrix of log(phi0) pre-estimated}

\item{model}{NB dispersion model name}

\item{method}{estimation method for dispersion to be passed to \code{estimate.dispersion.var}, either "ML" or "MAPL" (default is "ML")}

\item{hessian}{(logical) whether to calculate the Hessian matrix (default is FALSE)}

\item{refit}{(logical) whether to refit the underlying NB dispersion model (default is TRUE)}
}
\value{
an object from \code{estimate.dispersion.var} (if refit) or \code{optimize.il.prior.var} (if not refit)
}
\description{
Estimate sigma on simulated datasets
}
\details{
This function is used for checking the empirical Bayes estimation of dispersion noise sigma on simulated datasets.
Two options are available, either to refit the dispersion model being studied, or do not refit (use true dispersions).
}
\author{
Yanming Di, Gu Mi
}

