% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim.fdr.check2}
\alias{sim.fdr.check2}
\title{Simulate datasets (from real data) to check FDR controls for different NB dispersion models}
\usage{
sim.fdr.check2(model = "Genewise", dt = "mouse", sub.col = 1:6,
  disp.vec = "Quadratic", init.fit = FALSE, m = 5000, n = 6,
  s = 1.5e+07, perc.de = NULL, fc.vec = NULL, noise.vec = 1.05,
  seed.vec = NULL, out.path = path.o, FDR = 0.1)
}
\arguments{
\item{model}{NB dispersion model name}

\item{dt}{data name}

\item{sub.col}{subset of columns}

\item{disp.vec}{dispersion name}

\item{init.fit}{whether to have an initial fit for the dispersion model}

\item{m}{number of genes}

\item{n}{number of samples}

\item{s}{library size}

\item{perc.de}{percentage of DE genes}

\item{fc.vec}{fold change}

\item{noise.vec}{noise}

\item{seed.vec}{random number generator seed for reproducibility}

\item{out.path}{output directory}

\item{FDR}{nominal FDR threshold specified below which we call a DE gene}
}
\value{
a data frame with variables to be used for plotting
}
\description{
We simulate read counts of two groups from a real dataset. The (log) mean is determined by normal distribution,
and the percentage of DE genes can be specified, with half up- and half down-regulated. Nominal FDR can be
specified, and we calculate the (empirical) true FDR and then compare.
}
\author{
Gu Mi
}

