% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.model.check}
\alias{sim.model.check}
\title{Fitting dispersion models and getting p-values for the power-robustness simulations}
\usage{
sim.model.check(sim.data.obj, model = NULL)
}
\arguments{
\item{sim.data.obj}{output from sim.data.prepare()}

\item{model}{NB dispersion model to be fitted, including: "Common", "NBP", "NBQ", "NBS", "Trended", "Genewise", "Tagwise-Common",
"Tagwise-Trend", "QL", "QLShrink", "QLSpline" and "Genewise-HOA"}
}
\value{
a list of two elements:
(1) dispersion model name ("model");
(2) DE test p-value, used for calculating TPR-FDR curve  ("pval.y")
}
\description{
Fitting dispersion models and getting p-values for the power-robustness simulations
}
\details{
We use the normalization method discussed in Anders and Huber (2010) for all dispersion models.
For QL and its variants, we use the upper 75% quantile approach as suggested in the package manual.
}
\author{
Gu Mi
}

